@@ -74,6 +74,9 @@ exports.generateTemplate = function(mailType, valueArray) {
   } else if (mailType == "Edit_Profile") {
     templateName = "editProfile.html";
     mailSubject = "Profile edited on Worktez";
+  } else if (mailType == "Demo_Request") {
+    templateName = "demoRequest.html";
+    mailSubject = "New Demo Request - Worktez";
   }
   const promise = getTemplate(templateName).then((data) => {
     if (templateName == "comment.html") {
@@ -146,7 +149,13 @@ exports.generateTemplate = function(mailType, valueArray) {
       data = data.replace("$recipientName$", valueArray[2]);
       data = data.replace("$name$", valueArray[1]);
       data = data.replace("$teamName$", valueArray[0]);
-      data = data.replace("https://worktez.com/verifyUser/$$$$", "https://worktez.com/verifyUser/"+valueArray[3]+ "/" + valueArray[0]+ "/" + valueArray[4]+"/"+valueArray[2]);
+      data = data.replace("$link$", '"'+"https://worktez.com/verifyUser/"+valueArray[3]+ "/" + valueArray[0]+ "/" + valueArray[4]+"/"+valueArray[2]) + '"';
+    } else if (templateName == "demoRequest.html") {
+      mailSubject = "New Demo Request - Worktez";
+      data = data.replace("$userName$", valueArray[0]);
+      data = data.replace("$userEmail$", valueArray[1]);
+      data = data.replace("$userContact$", valueArray[2]);
+      data = data.replace("$orgName$", valueArray[3]);
     }
     subjectMessage = mailSubject;
     message.push(subjectMessage);
