@@ -75,7 +75,7 @@ export class TaskDetailsComponent implements OnInit {
 
   gotTaskData: boolean=false;
   showLoader: boolean=false;
-  
+ 
 
   activityData: Activity[]
   linkData: Link[]
@@ -305,7 +305,7 @@ export class TaskDetailsComponent implements OnInit {
     const callable = this.functions.httpsCallable( 'tasks/log' );
     const appKey = this.backendService.getOrganizationAppKey();
 
-    callable( {AppKey: appKey, SprintNumber: this.task.SprintNumber, LogTaskId: this.task.Id, LogHours: 0, LogWorkDone: this.task.WorkDone, LogWorkStatus: "Ready to start", LogWorkComment: "Reopening", Date: this.creationDate, Time: this.time, Uid: this.authService.user.uid } ).subscribe({
+    callable( {AppKey: appKey, SprintNumber: this.task.SprintNumber, LogTaskId: this.task.Id, LogHours: 0, LogWorkDone: 0, LogWorkStatus: "Ready to start", LogWorkComment: "Reopening", Date: this.creationDate, Time: this.time, Uid: this.authService.user.uid } ).subscribe({
       next: (data) => {
         this.getTaskPageData();
         this.showLoader = false;
